made add-cron safer

+++ add-cron.sh
@@ -1,9 +1,16 @@
 #!/bin/bash
 
-curr_date=$(date +%m%d%H%M)
+crontab=$(crontab -l)
+search="rig-temp-monitor"
 
-crontab -l > mycron
-crontab -l > /cronBackups/mycronBACKUP$curr_date
-echo "@reboot sleep 120 && python3 /rig-temp-monitor/rig-temp-monitor.py" >> mycron
-crontab mycron
-rm mycron
+if [[ "$crontab" == *"$search"* ]]; then
+    echo "It's there."
+else
+    echo "It's not there. Adding it to cron"
+    curr_date=$(date +%m%d%H%M)
+    crontab -l > mycron
+    crontab -l > mycronBACKUP$curr_date
+    echo "@reboot python3 /rig-temp-monitor/rig-temp-monitor.py" >> mycron
+    crontab mycron
+    rm mycron
+fi
